[CilkSanitizer] Add support for static race detection to make CilkSanitizer's instrumentation more precise.
At a high level, CilkSanitizer now instruments a module in several
phases:

1) Examine all instructions -- specifically, memory accesses and call
sites -- that could result in a race.

2) Use DependenceAnalysis and AliasAnalysis to perform static race
detection to try to prove which of these instructions cannot
participate in a race.

3) Propagate information about which functions might contain racing
instructions internally to all functions in the module, to determine
which callsites must be instrumented and which can be suppressed from
performing CilkSan updates.

4) Insert instrumentation around all potentially racing instructions
and other necessary IR objects, e.g., Tapir instructions, memory
allocation, and function entry and exit.