Merge with 16.x
This includes all the work by Pat McCormick <pat@lanl.gov> to add support for
AMDGPU.s This also includes an overhaul of the kitsune/experiments directory
and everything in it.

There is still some work to be done - for instance, getting rid of the legacy
pass manager for everything except backend code generation, but this should
be in a functional state for now.