[kitsune] Fix peculiar build-time behavior
Because of the way the build system was set up, the targets and kitrt would be
"installed" at build time i.e. when running ninja/make as opposed to
ninja install/make install. This fixes that behavior and only installs during
ninja build/ninja install. When building, there will be messages that suggest
that those targets are actually being installed, but they are being "installed"
to a subdirectory within the build directory.