dev/18.x linking error fixes (#49)
clang/lib/Frontend has a call to a Tapir << operator, which means it has
to be linked against TapirOpts.

There's unguarded checks to Value::dump in HipABI, which is disabled for
release builds, so I've replaced them with LLVM_DEBUG calls. Note this
means you don't have that output before assertion failure for release
builds. Other options if that's important.