[kitsune] Add comment describing why cheetah is checked out from a different repo
There is a bug caused by EarlyCSE stripping away the types of some Cilk objects
when building libopencilk-abi.bc. This breaks the OpenCilk Tapir target. A
temporary workaround has been added to Cheetah and this fork is in my repo from
which Kitsune will checkout Cheetah if the OpenCilk target is enabled. This is
a temporary measure and should be changed once upstream addresses the issue.