[kitsune] Support for Kitsune's custom memory allocator
This is a compiler-specific allocator that lowers to an appropriate memory
allocation routine depending on the tapir target being used. This is relevant
particularly for the GPU targets where the appropriate UVM (Unified Virtual
Memory) from either libcuda or libamdhip64 will be called. These are not called
directly. Instead the lowering uses calls to the appropriate wrapper from
libkitrt.

This consists of special builtins in the frontend that are now recognized by
Kitsune. These are lowered to Kitsune-specific LLVM intrinsics before handing
off to the middle-end. A middle end pass was added that lowers the intrinsic to
an approriate call to the Kitsune-runtime.

On the user-facing side of things, a templated kitsune::mobile_ptr<T> type is
provided that is syntactically similar to a std::unique_ptr<T> (semantically,
however, it is, obviously, very different). The experiments were all updated to
use these instead of the existing calls to the runtime. Currently, kitsune.h
will need to be #include'ed in order to use these. This is probably what we want
to encourage anyway. Even if we ever add support for this templated class into
the compiler itself, we will still want kitsune.h to be used to allow the code
to be compiled on other compilers.

Eventually, a [[kitsune::mobile]] attribute will be added that will allow the
backend to identify pointers that were allocated in UVM.