[kitsune] Add kit-config tool (#59)
Add a kit-config tool that is similar, in principle, to llvm-config. This can
be used to query the Kitsune build to determine which Tapir targets have been
built, whether Kokkos mode has been enabled etc. This is an initial
implementation and currently does not do much more than query the Tapir targets.