[kitsune] Add support for the kitsune::mobile attribute
This includes a couple of different spellings. A lot of the frontend interface
that uses kitsune_mobile_alloc and kitsune_mobile_free was also changed. Some
semantic checks have been added as well as tests. Several of the tests had to
be updated as a result.

Although not strictly related, the tests for the memory access attributes were
moved and some related code was cleaned up along the way.