[kitsune] Fix for libcuda in non-standard paths
For some reason, cmake does not look in /usr/lib64 when looking for libraries.
It also seems that, on some systems at least, that path is not on ld.conf
either. That, or the presence of a different ld.conf can confuse things. At any
rate, this manifests as the inability to find libcuda - either at configure time
or at runtime.

The cmake part of this is hack-y and should be fixed. It simply hardcodes the
paths to /usr/lib64 and /usr/local/lib64 as places to look when finding a
library. Surely there must be a better way to get cmake to look in those places,
but I haven't found any (I suppose I could have looked harder, but we'll just
defer it for now).

On the driver side, the paths to the directory containing libcuda are now
explicitly passed to the linker with -L and -rpath.