[kitsune] Require LLVM_BINUTILS_INCDIR to be set at configure time
This is required to build the gold plugin which is needed for LTO. Since we
force LLD to build, we might as well force this to built as well. Otherwise,
user code will fail to compile which is, obviously, not a good experience. If
we ever switch to doing everything at link-time, we will need this anyway.