[kitsune] Remove legacy Tapir passes
Removes most of the Tapir passes for the legacy pass manager. Not all have been
removed. The major one that is retained is TaskInfoWrapperPass because that is
used by a large number of standard LLVM passes and removing it from all of them
is somewhat fiddly. There may still be others that I have missed, but for now,
it's a start.