[kitsune] Add README for old kitsune-tests directory
Add a README to the kitsune-tests directory at the top-level of the repo. This
directory needs to be removed eventually, but the contents will have to be
checked to see if anything is worth retaining first. Until then, the README
should (hopefully) stop people from paying any heed to that directory.