[kitsune] Add pass to strip kitsune's address spaces from pointers
Kitsune uses address spaces to indicate pointers to mobile buffers. These
address spaces cause issue with some tapir targets, notably hip. This adds a
pass to be scheduled just before the tapir passes are run which mutates the
types in the module to replace pointers in kitsune's address space(s) with ones
in the default address space.