[kitsune] Set FPOpFusionMode in the GPU ABI's TargetOptions correctly
Correctly propagate the value of -ffp-contract down to the GPU ABI's. Also set
the value correctly from the -fp-contract flag when using opt.

Also cleanup the llvm lit configuration for the tests. This should really
belong in a separate commit, but all this will be squashed into a single, giant
commit when merging with LLVM 20.x anyway, so why bother?