[kitsune] Add KITSUNE_GCC_INSTALL_DIR configure-time option
This obviates the need to specify --gcc-install-dir, either in a config file,
or explicitly when invoking the compiler. If not provided, clang willlook for a
candidate GCC installation in the usual places. If the  --gcc-install-dir
command line option is used, or if it is set in a config file, that will be
used instead of value provided at configure time.