[kitsune] Bugfix is Kokkos code generation
MDRangePolicy's in Kokkos allow the first argument to be an iteration space.
However, we only handle the case where the first two arguments are bounds. For
now, just fall back to standard C++ lowering if the first argument is an
iteration space. Also enable the assertion that ensures that we have correctly
cleaned up the tapir lowering stack in clang.