[kitsune] Fix sysroot handling
On systems that require a sysroot argument for clang to work correctly, ensure
that it gets one when building the runtimes (both third-party, including Kokkos
and kitrt itself). There is a temporary workaround where cmake's check of the
compiler is disabled when building these (since they are built as external
projects with the clang that was just built, cmake will automatically run these
to ensure that the comiler we have told it to build with actually works). A
better solution may be to use config files with --sysroot set in them. This will
mimic the way clang would have to be used on such systems since they will,
inevitably, require a configuration file which provides --sysroot.