[kitsune] Remove all LLD customizations
All tapir-specific command line arguments are now passed to LLD as -mllvm
options. This is not ideal because this is essentially treating LLD as if it
were opt, but this spares us from having to update arguments for every LLD
tool (ELF, COFF, MachO) whenever we add an option to either kitsune's frontend
or as mllvm arguments in the tapir targets.