[kitsune] Refactor to share options objects, create a TapirTargetAnalysis pass
The KitsuneOptions object must be shared between clang and flang. This moves
the definition of this class to llvm/Frontend which is intended to be a place
where code to be shared between clang and flang lives. The object was also
originally stashed inside clang's LangOptions object. This was always intended
to be temporary since the Kitsune-specific options include language, codegen
and target options. As part of the cleanup, this was moved to the
CompileInstance object where it now lives at the "top-level" alongside
LangOptions, CodeGenOptions etc.

The TapirTargetOptions object was also moved to LLVM/Frontend. The
PipelineTuningOptions object was extended to accept a TapirTargetOptions. This
is now used to pass options to the tapir targets. The LowerTapir parameter
was removed from the pass pipeline builder functions. These now use the
TapirTargetOptions object from the PipelineTuningOptions object to determine
whether the tapir lowering passes should be enabled.

The TapirTarget is removed from TargetLibraryInfo which is returned by the
TargetLibraryAnalysis function pass. This was replaced with a new
TapirTargetAnalysis module pass. This new pass also contains the
TapirTargetOptions object. The instance is created when the module analysis
passes are registered.

The LLD tools for the various object file formats were modified to construct
a TapirTargetOptions object from the command line options passed to it. This
should be the only change that needs to be made to LLD for LTO support. The
kitsune tests were updated to include more fine-grainted tests to ensure that
LTO behaves as expected