[kitsune] Overhaul command line options in front and middle-ends
Several command line options from the cuda and hip ABI's are promoted to
clang/flang. The values of these options are added to the TapirTargetOptions
object. The tapir targets were updated to use these options. For example,
HipABI had a command line option to set the xnack attribute which is now part
of clang's options.

In some cases, if an equivalent option was already present in the clang/flang
frontend, that was used. For example, the -mcode-object-version= option can be
used to set the ABI version of AMDGPU object files. In such cases, code was
added to the Kitsune-specific parts of the toolchain to handle these options.

A lot of the work to compute GPU-specific options has been moved out of the
ABI's and into clang where the corresponding GPU toolchain is used. For example,
the AMD toolchain is used to determine the set of bitcode files to link into
the kernel module. These are then passed to the tapir targets via -cc1 [-fc1]
options.

When using the GPU tapir targets, the system will be queried for a GPU. If a
unique GPU is found, code is generated for exactly that architecture. If a GPU
is not found, code is generated for a default GPU architecture.

For CudaABI, the runtime bitcode file is not computed at configure-time.
Instead, the file to use is queried from the NVPTX toolchain.