[kitsune] Enable Kitsune-specific command line option processing and pass pipeline setup in Flang
Refactor the code that handles command line options so it can be shared between
clang and flang. For now, the refactor relies on clangBasic even though the
flang developers would like to remove the dependence of flang on clangBasic.
This is ok for now because it is not yet clear how flang will disentangle itself
from it. When they do, we should be able to follow their lead relatively easily.

Add hooks in flang to parse the Kitsune-specific command line arguments. Also
setup the tapir target options so the tapir passes are enabled if the --tapir
flag is passed on the frontend. There is still no support to actually lower any
Fortran constructs to tapir instructions. But we can still test that the
command line options are propagated to the tapir targets via the options object.

The Fortran tests were enabled. Some extensive reorganizing of the tests was
carried out so there is some semblance of consistency in what gets tested where.
Some LTO tests were also updated, so we know that the options do make their
way to the linker when compiling Fortran with -flto.