[kitsune] Refactor to use the optimization level from TapirTargetOptions
Currently, the optimization level is saved in the TapirTarget by threading it
through the TapirToTarget and LoopSpawningTI passes. This has been removed and
the optimization level saved in the TapirTargetOptions object is used instead.
This required some changes to the way the TapirTargetAnalysis pass is created
since, with tools like opt, we have to wait until the pipeline is parsed before
we can be sure what the optimization level is.

Some code that was duplicated in the cuda and hip tapir targets has been
shared.