[kitsune] Rename threads-per-block and max-threads-per-block frontend options
The new name is more consistent with the convention that requires options not
common to all tapir targets to have an element that indicates the tapir
target(s) to which it applies. In this case, it adds -gpu to the name to keep
it consistent with the --tapir-gpu-prefetch option which indicates that the
option is relevant to all GPU-centric tapir targets.