[kitsune/test] Fix LTO tests
Add a sysroot option to the LTO tests to ensure that they work correctly on
systems with non-standard sysroots. Also use llvm-dis when disassembling
from bitcode to human-readable LLVM assembly instead of using opt. Not only is
this likely to be faster, but more importantly, it makes it more obvious what
the test's run line is actually doing.