[kitsune] Fixes to allow ROCM installation in non-standard location
Fix some handling of ROCM in the build system. Also fix the toolchain
construction (for the cc1/fc1 call) to work with ROCM installations in
non-standard locations. The latter involves passing the path to the libdevice
bitcode directory explicitly to cc1 in addition to --rocm-path which was already
being passed.