[kitsune] Fix tests on AArch64 post merge with dev/20.x
In some cases, the tests needed to be updated to match upstream LLVM which
somehow got missed, others needed the -tail-dup-cbz=0 option and a few needed
to be updated because, presumably, the slightly different set of passes that
we run results in different output from what was expected.

In at least one case, the test has a bug because it will always fail if the
X86 target is not built. This really ought to be fixed upstream, but for now,
we just change it here.