[kitsune] Create and use intrinsics for kitsune runtime functions
Currently, kitsune's runtime is mainly used for the cuda and hip tapir targets.
It primarily wraps a lot of functionality provided by libcudart and libamdhip64
for NVIDIA and AMD GPU's respectively. The actual API is often similar between
the two. There are some functions exposed by kitsune's runtime that are only
relevant either NVIDIA or AMD GPU's, but not both.

The new kitsune-specific intrinsics cover both of these situations. In the
first case, the intrinsics take a tapir target id as an argument. In the second
case, the name of the tapir target to which the intrinsic applies is included in
the name of the intrinsic. The corresponding lowering pass that was added uses
the tapir target id to pick an appropriate runtime function to use.

The cuda and hip tapir targets have been modified to emit these intrinsics
instead of directly calling kitsune's runtime functions. While doing so, some
code was also cleaned up - primarily code that was written when pointers were
not opaque. The "TapirTargetFuncs" code has been removed since it did not seem
to be used in any meaningful way.

Kitsune's runtime functions are registered with TargetLibraryInfo. The lowering
pass uses much of the infrastruture that is provided to deal with these.

A KitsuneUtils file was added to Transforms/Utils. Some shared code could be
moved there, including from TapirGPUUtils, but that has not been done here.