[kitsune] Fix kitsune tests
The C++ frontends and clang were being used in the .f90 tests. Fix those to
use kitfc and flang respectively (sloppy copy-paste bugs). Also fix the test
that checks that we do not allow OpenMP offload and Tapir simultaneously. The
test now only emits the error that we are interested no other. Earlier, both
errors were being emitted, so the test would pass. This is cleaner.