[kitsune] Split CudaABI into separate passes
This is a major change to the way the GPU tapir targets operate. Rather than
do everything right up to generating a fat binary in a single step, this
splits out much that functionality into separate passes. The separate kernel
module into which the tapir loops have been extracted is in a global variable.
A number of passes operate on this "embedded bitcode". The passes are scheduled
in the main pass pipeline and purport to be module passes, though the only
change they make to the module is to update the initializer of the global
variables with the serialized updated bitcode. Similarly, a global variable
containing the fat binary is also created by CudaABI. This is not initialized
by CudaABI, or by any of the passes that operate on the embedded bitcode.
However, it is referenced where needed. A single passes was added to the
codegen pipeline which finally compiles the embedded bitcode modules and
populates the global variable with the fat binary.

These global variables are identified with Kitsune-specific metadata. Utilities
to process this metadata have been added.

The TapirTargetAnalysis pass was modified to be able to deal with multiple
tapir targets (via attributed tapir loops) if needed, although this code has
not been extensively tested - much as most of the multi-target support has not.

The verifier was updated to check the expected number of global variables
containing bitcode and the fat binary.

The split of CudaABI is not complete as of this commit and much of the code in
CudaABI has been commented out with the intention of cleaning it all up when
the entire process is complete and has been tested.