[kitsune] Add tool to extract embedded bitcode from a module
The tool will write the embedded bitcode modules to stdout. Options are provided
to write to file and to print/write both LLVM assembly and bitcode and to filter
the extracted embedded bitcode by the tapir target that created it. The
behavior of the tool is intentionally similar to opt.