[kitsune/tools] Add a module encoding utility
This utility takes an LLVM module as input (either from a file or stdin), embeds
the bitcode into an empty "host" module, then prints the "host" module to stdout
as LLVM assembly. This is helpful for test generation since we can add the
embedded bitcode global variable into a test.