[kitsune] Refactor HipABI
Split HipABI to use the same sequene of lowering steps as CudaABI i.e. a number
of separate passes. Along the way, a lot of the code in HipABI was cleaned up.
Much of the cleanup involves limitations imposed by Hip that are no longer
relevant, but in some cases, they were approaches that seemed to help mitigate
a number of other issues in the code. Some very serious bugs in the code
involving address spaces were fixed. The testing coverage for both the hip and
cuda tapir targets has been expanded considerably. A few issues with the cuda
tapir target have also been resolved.