[kitsune] Bugfixes and (temporary) embedded bitcode verifier
- Fix idempotency bugs in some of the Kitsune passes
- Fix bug with incorrect linkage when resolving libdevice functions
- Add pass to verify embedded bitcode.

The pass that verifies the bitcode should be removed in favor of verifying the
embedded bitcode in the main verifier. This has a direct impact on the planned
refactor of Kitsune to consolidate as much of the code as possible into its
own directory. For now, we keep the verifier around, but we will switch to
the refactor and remove it when the refactor is completed. The churn is annoying
but, given the enormous changes that have been taking place already, it is
relatively minor.