[kitsune] Major code reorganization
This is a major reorganization of the code base to consolidate as much as
possible of the Kitsune-specific code in one place. This includes a number of
changes:

  - All Kitsune-specific source code has been moved to llvm/lib/Kitsune

  - That directory contains several subdirectories that are organized in a
    manner similar to llvm/lib. For instance, Kitsune's analysis passes are now
    in llvm/lib/Kitsune/Analysis, support code is in llvm/lib/Kitsune/Support
    etc. The intention is to move this code to ${CMAKE_SOURCE_DIR}/kitsune/lib
    if possible.

  - All headers for Kitsune code has been moved to
    ${CMAKE_SOURCE_DIR}/kitsune/include/kitsune. Like llvm/include/llvm, this
    directory contains subdirectories that mirror the structure of
    llvm/lib/Kitsune.

  - The unittests have been moved and organized in a similar manner

Along the way, a number of changes were made to ensure eliminate dependency
cycles between the libraries. However, there still exists exactly one cycle
which cannot be removed without major reorgnaization. This is the cycle between
LLVMKitOpts and LLVMPasses which cannot be removed since one of the Kitsune
passes builds a pass pipeline to run optimization passes on embedded device
modules. This has been documented in llvm/CMakeLists.txt

Along the way, GlobalVariable attributes were added to replace the use of
metadata to identify global variables containing embedded bitcode and fat
binaries, as well as function attributes to identify kernel and device
functions.