[kitsune/test] Reorganize tests
Reorganize kitsune/test into a deeper heirarchy to ensure that we don't have
too many tests/directories at a given level. This is in preparation of further
reorganization where tests of Kitsune-specific passes that are currently in
llvm/test will be moved here.

The README.md was completely rewritten to reflect this new organization. Some
text was added to provide some guidance for developers.