[kitsune] Restore LLVM OptimizationLevel. Introduce Kitsune-specfic opt level
The previous commit moved LLVM's OptimizationLevel object out of libLLVMPasses
and into libLLVMSupport to avoid circular dependences between libLLVMKitsuneOpts
and libLLVMPasses. However, a major issue that was overlooked was that, with
compilers that support it, libLLVMSupport is compiled with
-Werror=global-constructors. Instances of LLVM's OptimizationLevel objects are
global, which obviously results in a compile-time error.

Since circular dependences are generally "bad things", a Kitsune-specific type
is created to represent the optimization level. This is utterly ridiculous, but
it is easier than trying to find a different home for the OptimizationLevel
objects.

Comments have been added to CMakeLists.txt explaining why circular dependences
have been introduced (in cases where they are not easily avoidable). Individual
libraries were added as appropriate to, in principle, support linking against
individual shared libraries.