[kitsune] Use a consistent naming scheme for passes
Passes that operate on embedded modules are prefixed with emb- while those that
operate on the host are prefixed with kit-. This is mostly just for consistency
and to easily identify passes that are kitsune-specific. The names of the passes
have also been shortened in some cases where it made sense to do so and/or
tweaked to better reflect what they actually do. The file names have also been
tweaked to reflect the pass name.

A Kitsune-specific TransformUtils library has been created