[kitsune] Tweak option to override optimization of embedded modules
Use -emb-O<N> instead of --emb-opt-level=. <N> can be 0, 1, 2 or 3 as well as
s or z. These correspond to LLVM's O0, O1, O2, O3, Os and Oz optimization
levels.

Also add a -emb-print-pipeline-passes option that prints the passes that will
be run on the embedded modules.

Add tests to check that these options are handled correctly.