[kitsune] Add option to print commandlines of external command invocations
When generating the fat binary, external commands are usually invoked such
as ptxas, or ld.lld. This adds a command line option to print those command
lines. Unlike the -### option in clang, this *does not* only print the
commands - rather it prints the commands, then executes them as normal.

Added some tests that uses this new option to check that sramecc and xnack are
handled correctly when codegen'ing the fat binary.