[kitsune] Add command line option to override the ptxas --opt-level value
When generating the fat binary, ptxas is invoked and given the --opt-level
option. The value of this option is the optimzation level set in the
TapirTargetOptions object. This adds a command line option that, if provided,
overrides the optimization in TapirTargetOptions.

The default optimization level set in TapirTargetOptions when using opt was
changed from O0 to O1. In practice, this shouldn't make a major difference
since, in most cases, we will use the tapir-lowering<N> or the
yet-to-be-implemented kitsune-lowering<N> meta passes where N will be the
optimization level. Setting this to -O1 is mainly useful when testing since,
at that point, we are more likely to run a single pass without using the
meta-passes.