[kitsune] Change Kitsune-specific global variable attribute handling
The kit_bc and kit_fb attributes no longer accept an integer value (the tapir
target). Instead, a new attribute, kit_tt was added whose value is the
integer representation of the tapir target that generated the attribute. The
kit_bc, kit_fb and kit_kernel_props attributes now require the kit_tt attribute
to be present. The verifier was updated to ensure that this is the case. Tests
were added for this as well and existing tests were updated.

kitmenc was updated to generate both an embedded bitcode and an embedded fat
binary global variable. This ensures that the module generated by kitemc will
actually pass when it is verified. The tool can also only be given a tapir
target that does generate embedded bitcode (currently either cuda or hip).
Passing any other tapir target will result in an error.