[kitsune] Fix command line handling in kitsune tools
Refactor the way command line options are handled in the kitsune tools. All
options are private to TapirTargetOptions, but utilities have been provided to
allow arguments to be selectively made visible and their descriptions
overridden. The tools themselves construct a TapirTargetOptions object to
"access" these shared arguments.

Some members of the TapirTargetOptions object were renamed to make them more
consistent with the rest of the code. Some defaults for the options were made
explicit and exposed as static members. These are not currently used.