[kitsune] Add option to override the codegen optimization level
The tests check both this and that the relocation model and code model (neither
of which can be overridden at this moment) are as expected. Doing so required
writing serialization routines for TargetMachine, TargetOptions and
MCTargetOptions. Command line options were added to dump these for the tests
to use. Serialization routines were also provided for several enums used in
CodeGen, MCTargetOptions and TargetOptions so that there would be a more useful
string value that is printed rather than just the integer value of the enum.
Some refactoring of the CodeGenFatBinaries pass and the TargetUtils functions
was carried to accommodate this additinoal flexibility.