[kitsune] Organize Kitsune's command line options into categories
The main Kitsune command-line options are defined in TapirTargetOptions.cpp and
are in the "Kitsune Options" category. Add the command line options that are
private to Kitsune's passes to a separate category, "Kitsune Developer Options".
Both categories are now public, although the command line options themselves are
still private to a file. Add a test to ensure that these two categories contain
(at least some) of the options that we expect.