[kitsune] Move command line options out of GenerateCtorCuda and GenerateCtorHip
Some command line options that control generation of Kitsune's global ctors
specific to the cuda and hip tapir targets were moved to GenerateCtorCuda.cpp
and GenerateHipCtor.cpp as part of the big refactor into separate passes. These
have been moved back to CudaABI.cpp and HipABI.cpp respectively. The options
themselves were exposed and declared extern in the kit-ctors pass.

These options are bundled into a single GenerateCtorOptions object in the
kit-ctors pass and passed to the cuda and hip ctor constructor functions. This
is now in keeping with the pattern used in other such passes.

These options were added to the appropriate category and the tests updated to
reflect this.

A test was added to check that the thread index calculation is as expected when
Y-axis launches are used.