[kitsune] Rename KitOpts library
libLLVMKitOpts has been renamed libLLVMKitTransforms. This is to keep things
consistent since all the other libraries in Kitsune share the name of the
directory in which they are declared. In LLVM, the library constructed from the
contents of the IR/ directory is named libLLVMCore, while the contents of
Transforms/ are compiled to libLLVMOpts. Since Kitsune already has a Core/
drectory that LLVM doesn't, we might as well be more consistent.