[kitsune] Add kit-lowering meta-pass
This pass is suitable for use in the -passes option recognized by several LLVM
tools such as opt. This runs Kitsune's pre-tapir passes, the tapir lowering
passes and Kitsune's post-tapir passes, in that order. The meta-pass accepts
an optimization level similar to the tapir-lowering meta pass. This level must
be O1 or greater.

The tapir-lowering meta-pass was modified to also require an optimization level
of O1 or higher.

Some documentation was cleaned up, a minor reorganization of tests and tests
for both the kit-lowering and tapir-lowering meta-passes have been added.