[kitsune] Raise an error if --tapir is provided to the frontend with -O0
This error is already raised when the 'kit-lowering' and 'tapir-lowering'
meta-passes are provided to opt. This extends that logic to kitcc, kit++ and
kitfc. However, there is a difference in the behavior of kitcc, kit++ and
kitfc. If the tapir target specified with --tapir is 'none', then an error is
*NOT* raised and compilation proceeds as normal. This is because --tapir=none
is very useful in generating "tapirized" IR. At some point, we may provide a
different mechanism to do this, at which point, this exception will be removed
and all the tapir targets will be treated the same.