[kitsune] Remove KitsuneOptions and move KitsuneOptions.h
KitsuneOptions was originally in llvm/Frontend/Driver and has been moved to
kitsune/Frontend. This is in keeping with keeping as much of Kitsune's code
within kitsune/ as possible. KitsuneOptions.cpp was removed since all the code
in it was moved to KitsuneOptionUtils.cpp.

The KitsuneOptions object was removed from the Preprocessor. There was no good
reason to have it there and it resulted in a number of changes to unittests
which had nothing to do with Kitsune. Most of these changes have been done so
they are back to being the same as upstream.