[kitsune] Fix kernel properties pass and some tests
The pass that computes the kernel properties was fixed to use the
kit_kernel_props attribute to identify the global variables containing kernel
properties and the value of the attribute to retrieve the kernel function.

This change in approach triggered a change in the expected values of several
tests. The earlier implementation would unconditionally insert a function
declaration in the module which would require some analyses to be recomputed.
This is not necessary any longer, and the tests were updated to reflect this.

The main test that checks for kernel properties was updated to check that the
values being computed by the pass were as expected.