[kitsune] Using --tapir=none should not add tapir.loop.target metadata
Currently, the value of --tapir is added as metadata to all tapir loops except
those that have a [[tapir::target]] attribute - for which the value of the
attribute will be set as the metadata. With this change, if --tapir=none, the
loop.target metadata will not be added to unattributed tapir loops.

KitsuneOptions was tweaked to be more consistent with the rest of the code and
use TTID instead of TapirTarget.

Assertions were added to ensure that the loop.target metadata matches the
primary tapir target. This is because multi-target support is currently broken.
When that is implemented correctly, the assertions can be removed.