[kitsune] Move intrinsic lowering and strip address space passes to codegen
The intrinsic lowering passes are usually run as part of codegen, not in the
middle end. Moving the passes there is in keeping with the rest of LLVM. The
pass to strip Kitsune's address spaces was modified to work with the other
passes and has also been moved to codegen.

The pass pipelines have been fixed to only run the Kitsune lowering and codegen
passes when the --tapir option has been given.

A number of other changes got included in this as well that are not strictly
related to the move of the passes to codegen, but cannot be committed
separately. These include:

  - Renaming the kitsune intrinsics to llvm.kit.* instead of llvm.kitsune.*

  - Renaming the library function enums to be more explicit about the direction
    of data movement. For example, copies from device to host were suffixed
    with _prefix, but are not suffixed with _dtoh. This is similar to the
    naming used in libcuda. Kitsune's runtime does not use this convention, but
    it may do so in the future.

  - Some minor refactoring to allow more code to be moved from llvm/lib to
    kitsune/lib. The code has not been moved yet, but the preparations are in
    place.

  - Some redudant tests were removed, while other tests were cleaned up so
    there are no extra attributes or other information that is not strictly
    necessary for the test.