[kitsune] Cleanup tests
Many tests contain unnecessary elements such as attributes on funcitons and
their arguments, compiler identification metadata and comments specifying the
predecessors of basic blocks. In the majority of cases, these are unnecessary
and distract from the properties of the code that are essential for the test.
These have been cleaned up to the extent possible.

Several tests on the embedded modules still run a full tapir lowering pipeline
followed by the actual pass being tested. This is unnecessary since what is
needed is just a global variable with the correct embedded module. However, this
would be very difficult to maintain since, if any change needed to be made to
the test, the embedded global - which is an encoded form of the bitcode would
need to be regenerated. Besides, it would be impossible to tell what the input
IR being tested was in this case without explicitly extracting the bitcode and
deserializing it. We could add the input IR as a comment in the test, but then
that would have to be kept in sync with the embedded module.