[kitsune] Rename the 'none' tapir target
The 'none' tapir target has been renamed 'nolo' to better indicate its actual
behavior i.e. it does not perform any lowering. The previous name was confusing
since 'none' suggested that tapir was not enabled, except that it was, in some
way.

The lower() method in TapirTargetOptions was removed and its uses were replaced
with an explicit comparison of the TTID with TTID::Nolo.