[kitsune] Move Kitsune pipeline construction to PipelineUtils
In addition to moving the code to construct the pre-tapir and post-tapir Kitsune
pipelines, the corresponding buildPreTapir buildPostTapir methods in PassBuilder
have been removed and a buildKitsuneLoweringPipeline() method has been added.
The codegen pipeline is only populated if tapir lowering is enabled, and this
is now checked by the populateKitCodeGenPasses instead of requiring callers to
check.

Some code in clang/lib/CodeGen/BackendUtil.cpp was cleaned up.

The PipelineUtils code was moved from kitsune/Core to kitsune/Passes to avoid
a circular dependence between libLLVMKitAnalysis (also libLLVMKitTransforms) and
libLLVMKitCore.