[kitsune] Make Kitsune's launch kernel intrinsic vararg
The arguments to be passed to the kernel are now passed as arguments to the
vararg intrinsic. The code to write each of these arguments to stack slots, then
bundling the pointers to the slots into an array to be passed to the runtime
function, and thence to the kernel, has been moved to the intrinsic lowering
codegen pass.

A bit of additional cleanup was performed on the Cuda and Hip ABI's.