[kitsune] Add a prefetching pass and move the prefetch generation code into it
The new prefetching pass runs very early in the post-tapir pipeline. Some
slight cleanup was carried out on both CudaABI and HipABI in addition to
removing the prefetch generation code.

The stream handling code was changed to explicitly get a stream from the
runtime. The stream is then used in the launch and prefetch calls. This is
different from the previous approach where the stream was "threaded" through the
prefetch and launch calls.

An intrinsic to get the thread stream was added, as were corresponding library
functions for Kitsune's runtime. The prefetch pass only inserts host-to-device
prefetches.