[kitsune] Move integer suffix from LoopOutlineProcessor to TapirTarget
An integer is used as a suffix to ensure that the functions containing the
outliend tapir loops have unique names. This integer was a static member of the
LoopOutlineProcessor's for the cuda and hip tapir targets. These have now been
moved into the corresponding tapir targets and are no longer static. The
complete name of the outlined kernel function can now be computed in a single
function call.