[kitsune] Better support to maintain debug information
Ensure that debug information is maintained when outlining tapir loops for the
GPU-centric tapir targets. The approach taken here is not really robust, but it
will do for now. For the purposes of debug info handling, the outliner pretends
as if the function is being outlined into the same module. This results in
"global" metadata nodes such as DICompileUnit, DIType etc. being shared between
the host and device modules. The verifier does not seem to have a problem with
this for the moment, but there is every chance that this will change. This has
not been tested with anything substantial yet, but for our relatively simple
examples, it seems to work. More details are provided in inline comments in the
code.

Tapir's outliner was modified to bring its organization more in line with the
cloning utilities in Transforms/Utils/CloneFunction.cpp. The makeshift support
for debug info handling was added there, so that will need to be updated when
we figure out how to handle debug information correctly (i.e. create dedicated
DICompileUnit, DIType etc. in the device module instead of sharing those nodes
with the host).