[kitsune] Add extension points into Kitsune's pass pipeline
These extension points allow a user to write a pass plugin that can be scheduled
before, after, or within Kitsune's pass pipelines. The extension points added
are:

  - PreTapirEarly: Runs early in the pre-tapir pipeline (before Tapir lowering)
  - PreTapirLate: Runs late in the pre-tapir pipeline
  - PostTapirEarly: Runs early in the post-tapir pipeline (after Tapir lowering)
  - PostTapirLate: Runs late in the post-tapir pipeline, before the global
                   ctors are generated
  - PostTapirLast: Runs at the end of the post-tapir pipeline, after the global
                   ctors have been generated

Currently, these are tested with a unit test. Ideally, we should add a pass
plugin in kitsune/examples similar to LLVM's Bye plugin and a regular lit test
to check that the passes are registered in the expected points in the pipeline.
However, the kitsune/examples directory is in dire need of some cleanup, so we
will defer adding a pass plugin until after we have cleaned it up. Until then,
a unittest will (have to) suffice.