[kitsune] Refactor attribute handling in kernel and device functions
Move some of the attribute handling code out of the tapir targets and into the
pass that prepares the device module for optimization and, eventually, code
generation. Only those attributes that have a direct bearing on GPU code
generation are retained in the tapir target. The attribute handling code that
was moved deals with targets (since the original target is always the host),
and those attributes that are not relevant for the GPU - for instance the
personality and uwtable attributes are removed because exceptions are not
currently supported on GPU's. These are also common to both kernel and device
functions, so they are shared by the kernel and device attribute handling code
in the prepare pass.